03 task and changed the name of the method

+++ CSharp-Technology-ADVANCED/Labs/08Generics-Lab/GenericArrayCreator/Program.cs
@@ -10,7 +10,7 @@ namespace GenericArrayCreator
             int[] integers = ArrayCreator.Create(110_122_222, 33);
             foreach (var @string in strings) Console.Write(@string + " ");
             Console.WriteLine();
-            foreach (long @integer in integers) Console.Write(@integer + " ");
+            foreach (int @integer in integers) Console.Write(@integer + " ");
             Console.WriteLine();
             for (int i = 0; i < strings.Length; i++) Console.Write (strings[i] + " ");
